Explicit file persistence
Different generate files need to exist in different ways. E.g., some need to be
committed to the repo so that other tools can see them (GH workflows, config for
GH apps, etc.). Others only need to exist for developers of the package, so they
can be uncommitted (and ignored) symlinks to the Nix store. And some are in
between – only needed locally, but with tooling that doesn’t like
symlinks (e.g., .gitignore).

This adds a `persistence` field to `file-type` that can be `store` (for
symlinks), `worktree` for uncommitted copy, or `repository` for committed copy.
I can see there being other types in future (like perhaps `ephemeral` for files
that only exist for the duration of certain commands), but this is currently sufficient.