Use bash-strict-mode
This also removes the `-b` option, which backs up instead of replaces files.
However, we expect to often be replacing files, so this doesn’t work for
us (also, this should generally only happen for committed files, so the danger
is less(?)).

Also, `-b` previously had a strictness issue when the backup extension wasn’t
set.