Add managed git config
This is the first successful example of using store persistence. Git hooks and
ignoreRevs aren’t linked into the working tree. The generated git config is also
not linked into the tree, but a reference is optionally added to the _real_ git
config, pointing to the rest of the config (including hooks and ignoreRevs) in
the Nix store.