Improve handling of profiles
Previously all project-manager projects shared a single profile, which was just
wrong and led to a bunch of confusing behaviors.

This improves the situation by having a profile per project-name. This still
isn’t quite right, as we should have one per working tree, but it’s a significant
improvement.

Down the road, I’m not sure how useful profiles are, other than the current
one (so we can reference it when it becomes the previous one). This is because
everything here should be managed via some VCS, so a previous generation
generally corresponds to a previous state of the repository, so why duplicate
that?

**NB**: This is incompatible with previous version of Project Manager.