Configure ShellCheck without generating a file
This is the model of a new feature in Project Manager – configuration without
files. It instead wraps the program to set up the environment or pass flags to
replicate the behavior of the config file.

This is controlled by three options:
- `wrapProgram` at the module level, which falls back to
- `project.wrapPrograms`, which falls back to (the opposite of)
- `project.commit-by-default`.

The last of those is because `commit-by-default` is generally a proxy for
non-Nix-using contributors, and those same contributors would also have trouble
with this feature.

This also gives more structure to the ShellCheck module, necessitated by having
to build either a config file or flags from it.