Invert flake inputs
This treats Flaky as the source of truth for inputs it provides. This currently
also means that Project Manager is upgraded to Nixpkgs 24.05, which also forced
a conversion of the documentation from DocBook and ASCII Doctor to
Markdown. (__NB__: Project Manager _already_ supported Nixpkgs 24.05 for
downstream projects, but it had still been built with Nixpkgs 23.11 itself.)