Fix unusable `init` subcommand
`project-manager` calls `nix flake metadata` to set `PROJECT_ROOT`. This had
been done before dispatching to a subcommand, and if `init` is used in a
directory without a flake, it will fail when it tries to set `PROJECT_ROOT`.

This now moves `PROJECT_ROOT` setting to a function and only does it for
subcommands that need it.

There are a couple other issues with `init` that followed this fix. E.g., the
`CONFIG_TEMPLATE` and `FLAKE_TEMPLATE` vars weren’t being substituted correctly.

There is still an outstanding issue with `init --switch` failing.