Fix `PROJECT_ROOT`
There were a few bugs around setting `PROJECT_ROOT`, such that it was
effectively always the directory `project-manager` was run from, which isn’t
ideal.

By looking up the root and setting the var directly in the `if` condition (as
recommended by [Shellcheck](https://www.shellcheck.net/wiki/SC2181)), we avoid
the fragility of suspending `trap` and disabling `-e`.

And now passing `--verbose` causes `PROJECT_ROOT` to be displayed.