Improve query parameter removal
`PROJECT_ROOT` sometimes has a query parameter attached to it. The previous
method of trying to remove it didn’t work, so this uses a simpler one that also
doesn’t rely on non-greedy matching.

We could do a more precise job with non-greedy matching, but `sed` doesn’t
support that, so for now just trim everything starting from the first `?`.